refactor: Gamers findPlayerByName 피드백 반영
