+++ src/main/java/blackjack/domain/BlackJackGame.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import blackjack.domain.Answer;
 import blackjack.domain.card.Deck;
@@ -8,7 +8,6 @@ import blackjack.domain.gamer.Player;
 import blackjack.domain.result.BlackJackReferee;
 
 import java.util.List;
-import java.util.function.UnaryOperator;
 
 public class BlackJackGame {
     private final Deck deck;
@@ -32,16 +31,12 @@ public class BlackJackGame {
         return gamers.getPlayers();
     }
 
-    public boolean isDrawPossible(String name, UnaryOperator<String> operator) {
-        return !gamers.canDraw(name) && Answer.from(operator.apply(name)).isYes();
+    public boolean isDrawPossible(Player player, String value) {
+        return !gamers.canDrawToPlayer(player) && Answer.from(value).isYes();
     }
 
-    public void distributeCardToPlayer(String name) {
-        gamers.distributeCardToPlayer(name, deck);
-    }
-
-    public Player findPlayerByName(String name) {
-        return gamers.findPlayerByName(name);
+    public void distributeCardToPlayer(Player player) {
+        player.addCard(deck.draw());
     }
 
     public int getDealerCardSize() {
