+++ src/test/java/blackjack/domain/BlackJackGameTest.java
@@ -1,9 +1,5 @@
 package blackjack.domain;
 
-import blackjack.BlackJackGame;
-import blackjack.domain.card.Card;
-import blackjack.domain.card.CardNumber;
-import blackjack.domain.card.CardShape;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.DisplayName;
@@ -55,24 +51,5 @@ class BlackJackGameTest {
             new BlackJackGame(List.of("name", "name"));
         }).isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("중복된 이름이 존재합니다.");
-    }
-
-    @Test
-    @DisplayName("플레이어 이름을 입력 받아, 플레이어를 반환한다.")
-    void findPlayerByName() {
-        BlackJackGame blackJackGame = new BlackJackGame(List.of("pobi", "jason"));
-        Player pobi = blackJackGame.findPlayerByName("pobi");
-        assertThat(pobi.isSameName("pobi")).isTrue();
-    }
-
-    @Test
-    @DisplayName("이름을 입력 받아, 해당 플레이어가 카드를 더 뽑을 수 있는 상태인지 확인한다.")
-    void isBurst() {
-        BlackJackGame blackJackGame = new BlackJackGame(List.of("pobi", "jason"));
-        Player pobi = blackJackGame.findPlayerByName("pobi");
-        pobi.addCard(new Card(CardShape.DIAMOND, CardNumber.KING));
-        pobi.addCard(new Card(CardShape.CLOVER, CardNumber.KING));
-        pobi.addCard(new Card(CardShape.HEART, CardNumber.KING));
-        assertThat(blackJackGame.isDrawPossible("pobi", s -> "y")).isEqualTo(false);
     }
 }
